requirements.txt
uvicorn~=0.28.0
sqlalchemy~=2.0.28
alembic~=1.13.1
fastapi~=0.110.0
jinja2
cryptography
starlette~=0.36.3
python-dotenv~=1.0.1
psycopg2
asyncpg
pydantic~=2.6.4